code/bot/requirements.txt
pytesseract
pillow
discord.py
requests
python-dotenv
